admin

+++ routers/todos.py
@@ -1,11 +1,11 @@
 from typing import Annotated
 from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
-
 from fastapi import APIRouter ,Depends, HTTPException , Path
 from starlette import status
 from Models import Todos
 from database import SessionLocal
+from .auth import get_current_user
 
 router=APIRouter()
 
@@ -17,6 +17,7 @@ def get_db():
         db.close()
 
 db_dependency= Annotated[Session, Depends(get_db)]
+user_dependency=Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str =Field(min_length=3)
@@ -25,33 +26,42 @@ class TodoRequest(BaseModel):
     complete: bool
 
 
-@router.get("/")
-async def read_all(db:db_dependency):
-    return db.query(Todos).all()
+@router.get("/", status_code=status.HTTP_200_OK)
+async def read_all(user: user_dependency,db:db_dependency):
+    return db.query(Todos).all().filter(Todos.owner_id== user.get('id')).all()
+
 
 @router.get("/todo/{todo_id}")
-async def read_todo(db:db_dependency, todo_id: int =Path(gt=0)):
-    todo_model=db.query(Todos).filter(Todos.id== todo_id).first()
+async def read_todo(user: user_dependency,db:db_dependency, todo_id: int =Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication falied')
+    todo_model=db.query(Todos).filter(Todos.id== todo_id)\
+        .filter(Todos.owner_id==user.get('id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail='Todo not found')
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db:db_dependency, todo_request: TodoRequest):
-    todo_model= Todos(**todo_request.dict())
+async def create_todo(user: user_dependency, db: db_dependency,
+                      todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication falied')
+    todo_model= Todos(**todo_request.dict(), owner_id=user.get('id'))
 
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency,
+async def update_todo(user: user_dependency,db:db_dependency,
                       todo_id : int,
                       todo_request: TodoRequest):
-    todo_model=db.query(Todos).filter(Todos.id== todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication falied')
+    todo_model=db.query(Todos).filter(Todos.id== todo_id)\
+        .filterfirst(Todos.owner_id ==user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
 
-    todo_model.id=todo_request.id
     todo_model.title=todo_request.title
     todo_model.description=todo_request.description
     todo_model.priority=todo_request.priority
@@ -61,10 +71,13 @@ async def update_todo(db:db_dependency,
     db.commit()
 
 @router.delete("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def create_todo(db:db_dependency, todo_id: int =Path(gt=0)):
-    todo_model= db.query(Todos).filter(Todos.id == todo_id).first()
+async def create_todo(user: user_dependency,db:db_dependency, todo_id: int =Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication falied')
+    todo_model= db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id ==user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id) .filter(Todos.owner_id ==user.get('id')).delete()
 
     db.commit()
